デッドリフト記録機能の実装

+++ db/migrate/20201211065754_create_bench_press_weight_records.rb
@@ -1,7 +1,7 @@
 class CreateBenchPressWeightRecords < ActiveRecord::Migration[6.0]
   def change
     create_table :bench_press_weight_records do |t|
-      t.integer    :bench_press_weight, null: false
+      t.float      :bench_press_weight, null: false
       t.date       :bench_press_day,    null: false
       t.references :user,               foreign_key: true
       t.timestamps
